agentic/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
google-generativeai>=0.3.2
langchain>=0.1.20
pymongo>=4.6.0
bson>=0.5.10
